% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p2type1}
\alias{p2type1}
\title{p2type1}
\usage{
p2type1(t, dom, alpha1, beta1, nu, alpha2, beta2, ancestors = 1L,
  threads = 1L)
}
\arguments{
\item{t}{time to run process}

\item{dom}{Size of domain of pdf to calculate. p2type1 calculates
values of P_{i,j}(x,y) for x,y = {0, 1, ..., dom - 1}. Larger
time requires higher values of dom, but slows down computation
since a dom x dom matrix is created.}

\item{alpha1}{type 1 split rate}

\item{beta1}{type 1 death rate}

\item{nu}{type 1 mutation rate (Note: not a probability!)}

\item{alpha2}{type 2 split rate}

\item{beta2}{type 2 death rate}
}
\value{
numeric matrix with (i,j) element representing P_{x,y}(x-1,j-1) since
  (0,0) is included.
}
\description{
p2type1 generates the probability distribution function P_{i,j}(x,y) for
a 2-type branching process with a single type 1 ancestor. Type 1
individuals split with rate alpha1, die with rate beta1 and mutate
to form one type 2 daughter with rate nu. Type 2 individuals
split with rate alpha2 and die with rate beta2.
}
