% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2type.R
\name{r2type}
\alias{r2type}
\title{r2type}
\usage{
r2type(N, pdf, ancestors = 1)
}
\arguments{
\item{N}{number of samples to generate}

\item{pdf}{pdf matrix generated from p2type}
}
\value{
matrix of randomly generated 2-dimensional values
}
\description{
r2type generates random numbers from a 2D distribution where probabilities
are described in a matrix. The function samples by running through the
matrix diagonally from (0,0).
}
\details{
Use this method after generating a distribution when you want to refer
to that distribution multiple times since the distribution is stored in
memory.
}
