% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2type.R
\name{rbdm}
\alias{rbdm}
\title{rbdm}
\usage{
rbdm(t, N, alpha1, beta1, nu, alpha2, beta2, ancestors = 1, model = 2)
}
\arguments{
\item{t}{time to simulate}

\item{N}{number of samples to generate}

\item{alpha1}{type 1 split rate}

\item{beta1}{type 1 death rate}

\item{nu}{type 1 mutation rate (Note: not a probability!)}

\item{alpha2}{type 2 split rate}

\item{beta2}{type 2 death rate}

\item{ancestors}{number of ancestors}

\item{model}{which model to simulate (1 = mutation without split, 2 = mutation with split)}
}
\value{
matrix of randomly generated 2-dimensional values
}
\description{
rbdm generates a 2-type birth-death-mutation process given the rate parameters
and returns basic summary statistics, the distribution function used, and the
simulated data.
}
\details{
Use this to generate many complete birth-death-mutation process
}
