% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2type.R
\name{ssa}
\alias{ssa}
\title{ssa}
\usage{
ssa(N = 1, t, alpha1, beta1, nu, alpha2, beta2, ancestors = c(1, 0),
  model = 2)
}
\arguments{
\item{N}{number of runs to generate}

\item{t}{time of simulation}

\item{alpha1}{type 1 birth rate}

\item{beta1}{type 1 death rate}

\item{nu}{type 1 mutation rate (model 1 generates a single type 2 and model 2 generates a type 1 and type 2)}

\item{alpha2}{type 2 birth rate}

\item{beta2}{type 2 death rate}

\item{ancestors}{2-element vector of number of ancestors}

\item{model}{1 or 2 for the model used}
}
\value{
2-element vector containing the number of type 1 and type 2 individuals
}
\description{
ssa generates multiple runs from the Stochastic Simulation Algorithm
under model 1 or 2. This can be used to compare to a distribution created in
the package.
}
